code made modular, file renamed
